Extract HiddenPatchNode
