+++ web/src/enterprise/campaigns/detail/patches/PatchNode.tsx
@@ -14,6 +14,7 @@ import { asError, isErrorLike } from '../../../../../../shared/src/util/errors'
 import { FileDiffConnection } from '../../../../components/diff/FileDiffConnection'
 import { FilteredConnectionQueryArgs } from '../../../../components/FilteredConnection'
 import { Observer } from 'rxjs'
+import { HiddenPatchNode } from './HiddenPatchNode'
 
 export interface PatchInterfaceNodeProps extends ThemeProps {
     node: PatchInterface
@@ -30,23 +31,6 @@ export const PatchInterfaceNode: React.FunctionComponent<PatchInterfaceNodeProps
     }
     return <HiddenPatchNode />
 }
-
-export const HiddenPatchNode: React.FunctionComponent<{}> = () => (
-    <li className="list-group-item e2e-changeset-node">
-        <div className="changeset-node__content changeset-node__content--no-collapse flex-fill">
-            <div className="d-flex align-items-center m-1 ml-2">
-                <div className="changeset-node__content flex-fill">
-                    <div className="d-flex flex-column">
-                        <div>
-                            <Octicon icon={Diff} className="icon-inline text-muted mr-2" />
-                            <strong className="text-muted">Patch in a private repository</strong>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </li>
-)
 
 export interface PatchNodeProps extends ThemeProps {
     node: IPatch
